requirements.txt
asyncio
discord
mcrcon
python-dotenv
requests
pytz
pytest-cov
codecov
youtube_dl
PyNaCl
openai
